Remove warnings and old deprecated code

+++ MSUScripter/Tools/StringExtensions.cs
@@ -7,7 +7,7 @@ public static class StringExtensions
 {
     public static string CleanString(this string str)
     {
-        return new string(str.Where(c => (int)c >= 0x1F).ToArray());
+        return new string(str.Where(c => c >= 0x1F).ToArray());
     }
 
     public static int GetUnicodeLength(this string str)
